requirements.txt
python-telegram-bot[job-queue]==20.1
schedule==1.1.0
requests==2.26.0
nest_asyncio==1.5.4
